Allow different config files for distinct machines
This commit adds the ability to install config files tailored exactly
for a predefined machine. The structure of InstallDotconfigs.sh script
makes it rather easy to add any number of other machines as well, if
needs arise. Currently two machines are being supported:

  * dotfiles_debian (for the Debian host at home)
  * dotfiles_work (for the Ubuntu system at work)

This commit updates the Debian configs. Ubuntu-related configs will be
updated in a future commit.